fix dexscreener price
